Ants are now viewable, somewhat
Ants can now be viewed F2 is pressed, but for some reason, whether or not
it is in the ant creation code or the ant view code, but even number ants
are not viewable.

Also moved the ant creation code from it's own function to inside the main
function. When it was independent the data was not accessable from the
view function.