Window scrolls
The window now scrolls but the text gets erased once it has been moved
off of the main view. Need to have it so that it finds out where it is
scrolled to and reprints all of the necessary information. Also, have
it so that you cannot scroll pas the last line of information.