can now select ants
changed the cell toggle to t and now with e you can select ant while on
top of them. the information for the currently selected ant will be
displayed at the bottom of the screen. If you press Shift + e it will
unselect an ant (this can be pressed anywhere)